Update restart script

+++ restart_experiment.py
@@ -6,32 +6,42 @@ from run_neat import sort_into_species
 from utils import PROJECT_ROOT, pluck
 
 if __name__ == '__main__':
-    problem_name = 'generate_tricolor'
-    db_file = '2016-12-04 18:09:12.299816.db'
-    from problems.generate_tricolor import CA_CONFIG, NEAT_CONFIG, FITNESS_F, PAIR_SELECTION_F
+    problem_name = 'generate_border'
+    db_file = '2016-11-23 13:51:24.771958.db'
+    from problems.generate_border import CA_CONFIG, NEAT_CONFIG, FITNESS_F, PAIR_SELECTION_F
 
     db_path = 'sqlite:///{}'.format(os.path.join(PROJECT_ROOT, 'problems', 'results', problem_name, db_file))
     db = get_db(db_path)
 
     session = db.Session()
 
-    for scenario_id, generation_n in [
-        (5, 5),
-        (49, 470),
-    ]:
+    for l in """""".strip().split('\n'):
+        l = l.strip()
+
+        if 'True' in l:
+            continue
+
+        scenario_id, generation_n, _ = l.split(' ')
+        scenario_id = int(scenario_id)
+        generation_n = int(generation_n)
+
+        if generation_n >= 500:
+            continue
+
+        print(scenario_id, generation_n)
+
+        generation_n -= 1
         genotypes = list(pluck(db.get_generation(scenario_id, generation=generation_n), 'genotype'))
 
         assert genotypes
 
         to_purge = db.get_individuals(scenario_id).filter(Individual.generation >= generation_n)
-        print('deleting', to_purge.count())
+        print(scenario_id, 'deleting', to_purge.count())
         to_purge.delete()
 
         session.commit()
 
         species = sort_into_species(genotypes)
-
-        assert len(species) > 1
 
         initialize_generation(
             db_path=db_path,
